\name{LFPStreamKeepersData Datasets}
\docType{data}
\alias{rawQualityData}
\alias{BIBI}
\title{LFPStreamKeepersData Datasets}
\description{
LFPStreamKeepersData Datasets.
}
\format{Comma Separated Value (csv) file}
\source{LFPStreamKeepersData System}
\references{
   LFPStreamKeepersData Users Guide
}
\keyword{ ~kwd1 }% __ONLY ONE__ keyword per line
